rename folders, add staff page, update header
