+++ client/src/landing_pages/Home/Content.js
@@ -118,7 +118,6 @@ export const Content = () => {
         <Typography className={classes.comingSoonTitle}>Coming Soon</Typography>
         <Typography className={classes.listOfCities}>
           {`
-Riverton
 Draper
 South Jordan
 West Jordan
