+++ client/src/landing_pages/Home/LearnMore.js
@@ -7,9 +7,8 @@ import ColorBar from '../../shared/ColorBar';
 const LearnMore = () => {
   return (
     <>
-      <ColorBar />
+      <Header logo />
       <Container>
-        <Header />
         <Content />
       </Container>
       <ColorBar />
