+++ client/src/landing_pages/Staff/Content.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function (){
+	return <div>Meet our staff</div>;
+}
