+++ client/src/landing_pages/Staff/Content.styles.js
@@ -1,0 +1,7 @@
+import {makeStyles} from '@material-ui/styles';
+
+const useStyles = makeStyles((theme) => ({
+  
+}));
+
+export default useStyles;
