+++ client/src/landing_pages/ThankYou/ThankYou.js
@@ -2,6 +2,7 @@ import React from 'react';
 
 import Container from '@material-ui/core/Container';
 
+import Header from 'shared/Header';
 import ColorBar from 'shared/ColorBar';
 import Footer from 'shared/Footer';
 import Content from './Content';
@@ -10,7 +11,7 @@ import logo from 'assets/refined-harmony.png';
 const ThankYou = () => {
   return (
     <>
-      <ColorBar />
+      <Header />
       <Container maxWidth='md' direction='column'>
         <Content />
         <Footer logo={logo} />
